+++ code/src/infra/graphics/point.h
@@ -14,48 +14,48 @@
 class Point {
 public:
 	Point();
-	Point(const glm::vec2& pnt);
-	Point(float x, float y);
+	Point(const glm::dvec2& pnt);
+	Point(double x, double y);
 
-	float GetX() const;
-	float GetY() const;
+	double GetX() const;
+	double GetY() const;
 	
-	void SetXY(float x, float y);
-	void SetX(float x);
-	void SetY(float y);
+	void SetXY(double x, double y);
+	void SetX(double x);
+	void SetY(double y);
 
-	glm::vec2 GetPoint() const;
+	glm::dvec2 GetPoint() const;
 
-	void SetPoint(const glm::vec2& point);
+	void SetPoint(const glm::dvec2& point);
 
 	// Calculates the angle (in RADIAN) of the vector which has the start point as this point and end point pnt
 	// Note that this not consider great circle calculation
-	float GetAngle(const Point& pnt) const;
+	double GetAngle(const Point& pnt) const;
 
 	// Return the direct distance of this point to given pnt
 	// Note that this not consider great circle calculation
-	float GetDistance(const Point& pnt) const;
+	double GetDistance(const Point& pnt) const;
 
 	// Shift the point in the given direction
 	// angle is in radian
-	Point ShiftPoint(float angleInRad, float distance) const;
+	Point ShiftPoint(double angleInRad, double distance) const;
 
-	const Point operator / (const float denumerator) const;
-	const Point operator * (const float value) const;
+	const Point operator / (const double denumerator) const;
+	const Point operator * (const double value) const;
 	const Point operator + (const Point& point) const;
 	const Point operator - (const Point& point) const;
 	const Point operator - () const;
-	float operator * (const Point& rhs) const;
+	double operator * (const Point& rhs) const;
 
 	Point& operator += (const Point& point);		
 	bool operator != (const Point& point);
 	bool operator == (const Point& point);
 		
-	glm::vec2* GetPointPtr();
+	glm::dvec2* GetPointPtr();
 
 protected:
 	//! The point data
-	glm::vec2 mPoint;
+	glm::dvec2 mPoint;
 };
 
 #endif	// INC_POINT_H
